requirements.txt
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
playwright==1.40.0
requests==2.31.0
numpy==1.24.3 